+++ components/src/lib/jschema/form_errors.js
@@ -116,9 +116,12 @@ function isPrimitiveTypeElement(element) {
  * For oneOf properties, AJV returns errors also for unselected elements.
  * This function returns true when such an error is found, in order to ignore them.
  * @param {object} error 
- * @param {import('../types/form').FormElement} element 
+ * @param {import('../types/form').FormElement} element
  */
 function ignoreUnselectedConditionalError(error, element) {
+  if (error.message === 'must match exactly one schema in oneOf') {
+    return true;
+  }
   if ('selectedItem' in element) {
     const { oneOf } = element.property;
     const selectedIndex = get(element.selectedIndex);
@@ -130,10 +133,19 @@ function ignoreUnselectedConditionalError(error, element) {
         const { allowedValue } = error.params;
         const selectedValue = element.discriminator.values[selectedIndex];
         if (allowedValue !== selectedValue) {
+          // ignore discriminator errors related to unselected discriminator values
           return true;
         }
       }
-      if (error.schemaPath.startsWith(`#/properties${element.path}/oneOf/${i}`)) {
+      if (element.selectedItem && error.instancePath === element.path
+        && !error.schemaPath.startsWith(element.selectedItem.schemaPath)) {
+        // ignore oneOf errors when schema path doesn't match
+        return true;
+      }
+      if (!element.selectedItem && error.instancePath === element.path
+        && error.schemaPath.startsWith(`${element.schemaPath}/oneOf/`)
+      ) {
+        // ignore oneOf errors when invalid discriminator value is selected
         return true;
       }
     }