+++ src/routes/v2/projects/[projectId]/workflows/[workflowId]/+page.svelte
@@ -661,7 +661,7 @@
 	async function onWorkflowTaskUpdated(updatedWft) {
 		selectedWorkflowTask = updatedWft;
 		workflow.task_list = workflow.task_list.map((t) => (t.id === updatedWft.id ? updatedWft : t));
-    await tick();
+		await tick();
 		argsChangesSaved = true;
 		setTimeout(() => {
 			argsChangesSaved = false;
@@ -892,9 +892,13 @@
 						tasks' execution statuses. Possible reasons include: (1) All jobs ran with an old
 						Fractal version (before v2.14, which has been available since May 2025), or (2) a
 						job-execution error took place before the first task started running.
-						<br>
+						<br />
 						If you need to review what jobs already ran, have a look at the
-						<a href="/v2/projects/{project.id}/workflows/{workflow.id}/jobs{selectedDataset ? '?dataset='+selectedDataset.id : ''}">job list</a>.
+						<a
+							href="/v2/projects/{project.id}/workflows/{workflow.id}/jobs{selectedDataset
+								? '?dataset=' + selectedDataset.id
+								: ''}">job list</a
+						>.
 					</div>
 				</div>
 			</div>
@@ -1127,6 +1131,8 @@
 													<ArgumentsSchema
 														bind:workflowTask={selectedWorkflowTask}
 														{onWorkflowTaskUpdated}
+														taskName={selectedHistoryRun.workflowtask_dump.alias ??
+															selectedHistoryRun.workflowtask_dump.task.name}
 														editable={false}
 														bind:this={argsSchemaForm}
 														argsSchemaNonParallel={selectedHistoryRun.args_schema_non_parallel}
@@ -1139,6 +1145,7 @@
 												<ArgumentsSchema
 													bind:workflowTask={selectedWorkflowTask}
 													{onWorkflowTaskUpdated}
+													taskName={selectedWorkflowTask.alias ?? selectedWorkflowTask.task.name}
 													editable={true}
 													bind:this={argsSchemaForm}
 													argsSchemaNonParallel={selectedWorkflowTask.task.args_schema_non_parallel}
@@ -1187,13 +1194,29 @@
 									{#if selectedWorkflowTask}
 										<WorkflowTaskInfoTab
 											projectId={workflow.project_id}
-											workflowTask={selectedWorkflowTask}
+											workflowTask={selectedHistoryRun
+												? {
+														...selectedWorkflowTask,
+														task: {
+															...selectedWorkflowTask.task,
+															version: selectedHistoryRun.workflowtask_dump.task.version,
+															type: selectedHistoryRun.workflowtask_dump.task.type,
+															command_non_parallel:
+																selectedHistoryRun.workflowtask_dump.task.command_non_parallel,
+															command_parallel:
+																selectedHistoryRun.workflowtask_dump.task.command_parallel
+														},
+														alias: selectedHistoryRun.workflowtask_dump.alias,
+														description: selectedHistoryRun.workflowtask_dump.description
+													}
+												: selectedWorkflowTask}
 											updateWorkflowTaskCallback={(up) => {
 												selectedWorkflowTask = up;
 												workflow.task_list = workflow.task_list.map((wt) =>
 													wt.id === up.id ? up : wt
 												);
 											}}
+											editable={!selectedHistoryRun}
 										/>
 									{/if}
 								</div>